requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#

alembic==0.8.6            # via flask-migrate
click==6.6                # via flask
Flask-Migrate==1.8.0
Flask-Script==2.0.5       # via flask-migrate
Flask-SQLAlchemy==2.1     # via flask-migrate
Flask==0.11.1             # via flask-migrate, flask-script, flask-sqlalchemy
itsdangerous==0.24        # via flask
Jinja2==2.8               # via flask
Mako==1.0.4               # via alembic
MarkupSafe==0.23          # via jinja2, mako
python-editor==1.0.1      # via alembic
SQLAlchemy==1.0.13        # via alembic, flask-sqlalchemy
Werkzeug==0.11.10         # via flask
